pyproject.toml
[tool.poetry]
name = "council-forge"
version = "0.1.0"
description = "A local multi-agent council simulation powered by LLMs."
authors = ["Antigravity <antigravity@deepmind.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
httpx = "^0.27.0"
loguru = "^0.7.2"
pydantic = "^2.6.0"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
